Address issue when regions are disabled by SCP
This addresses the issue when AWS regions are disabled by SCP. It is common for organisations to disable unused regions. Moving the code block which checks config recorder status inside "try...except" block ensures that the application does not error out if it comes across a region which is disable by SCP. The exception is gracefully handled and the for loop moves on to check the config recorder status for the next region in the list.